// web/src/services/proto/api.ts
export interface ApiItem {
    /**
     * @generated from protobuf field: string id = 1;
     */
    id: string;
    /**
     * @generated from protobuf field: string name = 2;
     */
    name: string;
    /**
     * @generated from protobuf field: float price = 3;
     */
    price: number;
    /**
     * @generated from protobuf field: string description = 4;
     */
    description: string;
    /**
     * @generated from protobuf field: string category = 5;
     */
    category: string;
    /**
     * @generated from protobuf field: bytes image = 6;
     */
    image: Uint8Array;
}
// ...
export interface ApiAddItemRequest {
    /**
     * @generated from protobuf field: string jwt = 1;
     */
    jwt: string;
    /**
     * @generated from protobuf field: string restId = 2;
     */
    restId: string;
    /**
     * @generated from protobuf field: string name = 3;
     */
    name: string;
    /**
     * @generated from protobuf field: float price = 4;
     */
    price: number;
    /**
     * @generated from protobuf field: string description = 5;
     */
    description: string;
    /**
     * @generated from protobuf field: string category = 6;
     */
    category: string;
    /**
     * @generated from protobuf field: bytes image = 7;
     */
    image: Uint8Array;
}
// ...
export interface ApiGetItemsRequest {
    /**
     * @generated from protobuf field: string restId = 1;
     */
    restId: string;
    /**
     * @generated from protobuf field: string category = 2;
     */
    category: string;
}
// ...
class ApiItem$Type extends MessageType<ApiItem> {
    constructor() {
        super("api.ApiItem", [
            { no: 1, name: "id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "price", kind: "scalar", T: 2 /*ScalarType.FLOAT*/ },
            { no: 4, name: "description", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 5, name: "category", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 6, name: "image", kind: "scalar", T: 12 /*ScalarType.BYTES*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message api.ApiItem
 */
export const ApiItem = new ApiItem$Type();
// ...
class ApiAddItemRequest$Type extends MessageType<ApiAddItemRequest> {
    constructor() {
        super("api.ApiAddItemRequest", [
            { no: 1, name: "jwt", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "restId", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "price", kind: "scalar", T: 2 /*ScalarType.FLOAT*/ },
            { no: 5, name: "description", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 6, name: "category", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 7, name: "image", kind: "scalar", T: 12 /*ScalarType.BYTES*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message api.ApiAddItemRequest
 */
export const ApiAddItemRequest = new ApiAddItemRequest$Type();
// ...
class ApiGetItemsRequest$Type extends MessageType<ApiGetItemsRequest> {
    constructor() {
        super("api.ApiGetItemsRequest", [
            { no: 1, name: "restId", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "category", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message api.ApiGetItemsRequest
 */
export const ApiGetItemsRequest = new ApiGetItemsRequest$Type();